+++ tests/main.rs
@@ -4,23 +4,12 @@ use predicates::prelude::*;
 use std::process::Command; // Run programs // Used for writing assertions
 
 #[test]
-fn invalid_input_file() -> Result<(), Box<dyn std::error::Error>> {
-    let mut cmd = Command::main_binary()?;
-    cmd.args(vec!["-i file/doesnt/exist.fx", "-g", "5mb", "-c", "20"]);
-    cmd.assert()
-        .failure()
-        .stderr(predicate::str::contains("is not fasta or fastq"));
-
-    Ok(())
-}
-
-#[test]
 fn input_file_doesnt_exist() -> Result<(), Box<dyn std::error::Error>> {
     let mut cmd = Command::main_binary()?;
     cmd.args(vec!["-i", "file/doesnt/exist.fa", "-g", "5mb", "-c", "20"]);
     cmd.assert()
         .failure()
-        .stderr(predicate::str::contains("No such file"));
+        .stderr(predicate::str::contains("does not exist"));
 
     Ok(())
 }
@@ -63,8 +52,8 @@ fn valid_inputs_raises_no_errors() -> Result<(), Box<dyn std::error::Error>> {
 }
 
 #[test]
-fn input_and_output_filetypes_different_raises_no_errors_but_warns(
-) -> Result<(), Box<dyn std::error::Error>> {
+fn input_and_output_filetypes_different_raises_no_errors() -> Result<(), Box<dyn std::error::Error>>
+{
     let mut cmd = Command::main_binary()?;
     cmd.args(vec![
         "-i",
@@ -77,9 +66,7 @@ fn input_and_output_filetypes_different_raises_no_errors_but_warns(
         "/tmp/out.fasta",
     ]);
 
-    cmd.assert()
-        .success()
-        .stderr(predicate::str::contains("file types are not the same"));
+    cmd.assert().success();
 
     Ok(())
 }